requirements.txt

### `requirements.txt`
```txt
tensorflow>=2.15.0
pandas
PyYAML
numpy
joblib